【ADD path.c】コマンドの実行ファイル検索ロジックを実装
